backend/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
PyJWT

